Added ADC functionality, fixed garbage serial input bug
We may need to tweak the ADC sample frequency, I can't tell without lab equipment if it's set to a high enough frequency for audio.

I've tested the ADC in with PWM out functionality using a potentiometer and LEDs and it's working as intended and demonstratable.

serial garbage issue was when the Due would restart/reset, there would be garbage in the input buffer which would trigger the response to show the Amplitude/Deadtime output string but as garbled mess. Now the output message won't show unless there's a valid change made.